Surafel.search and filter

+++ Rideshare.Application/Common/Dtos/RideRequests/SearchAndFilterDto.cs
@@ -5,7 +5,7 @@ namespace Rideshare.Application.Common.Dtos.RideRequests;
 public class SearchAndFilterDto
 {
     public Status? status {get;set;}
-    public int? fare {get;set;}
+    public double? fare {get;set;}
     public string? name {get;set;} 
     public string? phoneNumber{get;set;}
