+++ Rideshare.Application/Contracts/Persistence/IRideRequestRepository.cs
@@ -10,7 +10,7 @@ namespace Rideshare.Application.Contracts.Persistence;
 public interface IRideRequestRepository : IGenericRepository<RideRequest>
 {
     Task<Dictionary<string, int>> GetTop5Commuter();
-    Task<PaginatedResponse<RideRequest>> SearchByGivenParameter(int PageNumber, int PageSize, Status? status, int? fare, string? name, string? phoneNumber);
+    Task<PaginatedResponse<RideRequest>> SearchByGivenParameter(int PageNumber, int PageSize, Status? status, double? fare, string? name, string? phoneNumber);
     Task<Dictionary<int, int>> GetAllByGivenParameter(string? type, int? year, int? month);
     Task<Dictionary<string, Dictionary<int, int>>> GetAllByGivenStatus(string? type, int? year, int? month);
     Task<PaginatedResponse<RideRequest>> GetAllRequests(int PageNumber, int PageSize);
